v2 overfitting-mitigation pass + GDP-HMM adapter upgrades + rename to pipeline_adapter
Training (v2, runs/full_real_run_v2/):
- model.py: FC head (1024,512,256)->(512,256), dropout 0.2->0.4 (new ctor arg); 2.20M->1.54M params
- train.py: weight_decay 1e-4->1e-3; early stopping on a 5-epoch moving-average
  val loss; checkpoint selection on that moving average (noisy 50-patient val);
  SWA with hand-rolled BN recalibration; best-vs-SWA compared on val for the
  test eval; MPS device support
- dataset.py: real geometric augmentation (flips + ±15° rotation + ±8%
  translation via scipy.ndimage, nearest-interp so masks stay binary)
- Result: early-stopped epoch 41, best moving-avg val epoch 26. Test (50
  held-out): D2 MAE 3.58->2.74, D50 8.01->6.66, Dmean 5.68->4.88 Gy, DVH-curve
  dev 9.2%->7.8% -- improved on every metric vs v1, no overfitting.

Adapter (pipeline_adapter.py, was gdp_hmm_adapter.py):
- resolves the PTV/target key and prescription dose from meta_files/PTV_DICT.json
  (PTV_High.OPTName/PDose) instead of a guess-chain; validates the chosen OAR
  against Pat_Obj_DICT.json; graceful fallback for patients not in the metadata
- also adds the missing resample-to-fixed-in-plane-matrix step (GDP-HMM dose is
  pre-aligned so align_dose_to_ct is skipped, but that skip had also dropped the
  separate resample-to-cfg.target_matrix step, so slices from different patients
  couldn't be batched)
- DEFAULT_META_DIR points at <repo-root>/pipeline/meta_files/ after the repo reorg

preprocessing.py: origin-aware dose->CT alignment fix + PixelSpacing row/col
order fix (carried over from earlier uncommitted work).

+++ preprocessing.py
@@ -309,8 +309,8 @@ def build_slice_dataset(study: PatientStudy, cfg: PreprocessConfig,
             "patient_id": study.patient_id,
             "oar_name": oar_name,
             "slice_index": z,
-            "target_mask": target_slice.astype(np.float32),   # channel 1
-            "oar_mask": oar_slice.astype(np.float32),         # channel 2
+            "target_mask": target_slice.astype(np.uint8),      # channel 1 (binary; cast to float32 at __getitem__ time)
+            "oar_mask": oar_slice.astype(np.uint8),            # channel 2 (binary; cast to float32 at __getitem__ time)
             "dvh_label": dvh,                                  # [num_bins]
             "voxel_count": int(oar_slice.sum()),               # for weighted aggregation
         })
